many to many relationship support
Add support for many to many relationship and the related
seql helpers.

It's expected that many-to-many relationships will be implemented
with an intermediate "join table". In this case the join table
will hold a reference to the source and destination tables.

The helper syntax proposed is:

    (has-many-through :remote [:id
                               :intermediate/left-id
                               :intermediate/right-id
                               :remote/local-id])

Which will expend to the following EDN:

    {:type               :many-to-many
     :local-id           :local/id
     :intermediate       :intermediate
     :intermediate-left  :intermediate/left-id
     :intermediate-right :intermediate/right-id
     :remote-id          :remote/local-id
     :remote-entity      :remote}

Tests show-case usage. Note that no name rewriting is possible
through join tables as they are generally not directly queried.