Add support for SQL IN queries in conditions
Currently, passing more than one argument to conditions of type
:field causes an error. But we often need to match a field on
various values. It can be done server side (get all + filtering) but
this can cause performance issues.

This commit modifies the :field condition to support multiple
parameters. If multiple parameters are provided, the where clause will
be a `IN` on all parameters.