Deeper integration with spec to allow schema inference
After a couple of years of using seql a couple of annoyances
surfaced:

- It is quite common for the entities stored in databases to have
  specs in the form of `s/keys` specs, the reliance on short
  qualifiers forces its use in specs or drives consumers towards
  replicating specs which is cumbersome and error-prone
- Idents and field conditions are superfluous and can be inferred
  from the schema or query structure
- Transforms are largely superceded by nicer spec based mechanisms
  such as those provided by coax (https://github.com/exoscale/coax)

This patch addresses the above while still supporting the previous
schemas. A breaking change is introduced: the removal of both
transforms and compounds. Transforms are now performed according
to specs, note that since coax is used, facilities are provided
to add readers and writers to specs to preserve the ability
to serialize fields to JSON or EDN, for instance.

When looking up relations, empty collections will have their field
elided from the output entirely. Documentation has been adapted to
the new code.